Implement Droid Sans as the default font, PP stuff
Yeah, I finally decided to switch fonts. After picking 6 different fonts, Droid Sans won the challenge. Testing on Linux will be performed once I set up a VM.
Big changes to the prettyprint stuff - this also requires the latest PR in the master repository (which will probably be merged soon, so don't worry about it). Now the 'Textual version' button is placed on the right of the title instead of the bottom. This IS really better than the old one and the callback system makes it almost instant.
The main font-size has also been adjusted, according to the new font (which is a webfont, so no more 'I do not have that font because I am a linux *beep* :('.