Better HTML, JS refactoring, CSS adjustments
The PM counter's position is not determined by margins anymore.
A compressed post will be expanded automatically if an image is clicked in it.
Removed unnecessary comments in JS code, better indentation (4 spaces everywhere)
Fixed searchbar's position.
Fixed font sizes for Linux.
The template is being tested on Fedora 20 with Infinality, so that everyone can have a consistent experience with it.