Introducing the Preferences API
Everyone has different tastes and preferences. That's why there should be a way to let a user customize a theme in whatever way he wants. Now available: the Preferences API.
A special, protheme-only file has been added which implements the Preferences API and other protheme-exclusive functions.
For now only two different preferences have been written, because the API is still in its early state. Also included there's a draft of the CustomLangsAPI, which will come soon.
Using the Preferences API is not hard. Reading the code is recommended, but I'm providing a fast-how-to for lazy users.
First, load the preferences relative to your theme with YourNamespace.PreferencesAPI.load ('themename').
Then you can hook whatever section of the Preferences page you want. Read the code for more information about that.
Finally - you need to use the addSettings() function to any kind of setting.
The Preferences API is not yet completed, and that's why documentation has not been provided yet. However, it is fully working and stable. It is also as fast as possible, providing the users a perfect experience.
Prepare to the revolution.